refactor(scripts): extract the shared GitHub client into github_api
check-stars.py's rate-limit policy — a 60-second floor for secondary
limits, Retry-After and x-ratelimit-reset handling, a 600-second per-run
sleep budget and a give-up message that names the limit — was paid for in
two broken scheduled runs. A second script needs the same GitHub client,
and a second implementation of that policy would simply re-earn those
failures, so the client moves to scripts/github_api.py and check-stars.py
imports it.

The move is docstring-only: is_rate_limited, parse_retry_after,
rate_limit_reset_text, rate_limit_wait, spend_rate_limit_budget,
describe_rate_limit, github_request and RateLimitError are byte-identical
to their previous definitions.

Two additions come with it. post_to_matrix now passes a timeout, which
the inline webhook POST did not: a wedged webhook used to hold the runner
open until the Actions timeout. reset_rate_limit_budget exists so tests
cannot inherit each other's budget through the module global.

The rate-limit tests move to tests/test_github_api.py, where the code
they exercise now lives, and gain two cases: a dict body must cost
exactly one request (the Actions endpoints return an envelope, and
paginating it would walk thousands of runs), and post_to_matrix must set
a timeout.

tests/test_check_stars.py gains a regression test for the scrape path in
get_dependents. It calls parse_retry_after directly, because it scrapes
github.com HTML rather than the API, which made it the one caller the
extraction left behind; the break is only reachable through a transient
5xx, so nothing else provoked it.
